Refactor create user defined function
